backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
python-socketio==5.11.0
aiofiles==23.2.1
httpx==0.26.0
Pillow==10.2.0
boto3==1.34.34
google-generativeai==0.3.2
openai==1.10.0
geopy==2.4.1
numpy==1.26.3
pandas==2.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
